server/usersManager/requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.3
Flask-JWT-Extended==4.4.4
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.2.3
zipp==3.15.0
